move one math utils function
