+++ pkg/utils/math.go
@@ -220,3 +220,13 @@ func sign(n *big.Int) (int, big.Int) {
 	}
 	return 1, *new(big.Int).Set(n)
 }
+
+func SafeDiv(x, y *big.Int) (big.Int, error) {
+	if y.Cmp(big.NewInt(0)) == 0 {
+		return *big.NewInt(0), fmt.Errorf("division by zero")
+	}
+	if new(big.Int).Mod(x, y).Cmp(big.NewInt(0)) != 0 {
+		return *big.NewInt(0), fmt.Errorf("%v is not divisible by %v", x, y)
+	}
+	return *new(big.Int).Div(x, y), nil
+}
